# GPU ported and optimised find_center_360: >420x speedup