## Overview

This PR ports the `find_center_360` function to the GPU using cupy, and optimises the function on both CPU and GPU. It also adds appropriate tests and it uses a unified function name for both CPU and GPU.

## Performance (426x faster)

Based on the performance test in this PR, with normalisation, filtering, and overlap enabled. Measured on Diamond cluster, V100

* Starting performance CPU: **~3,070ms**
* Straight port to Cupy, using `get_array_module`: **~7,150ms**
* Final performance CPU: **~1,580ms**
* Final performance GPU: **~7.2ms** (426x faster than initial CPU, 993x faster than straight GPU port)

## Details

* The key item was to put the calculation of all the correlation metrics (with all possible position shifts) into a single kernel
* The Pearson correlation coefficient can in fact be calculated easily by hand, for every position in a different thread block
* Also removed a lot of temporary arrays and small operations, which also benefited the CPU implementation
* Apart from that, unified tests have been added and the CUDA code has been moved into its own folder, which makes development much easier and code better organised