@team-gpu 2023-03-07
> > I've refactored this PR as discussed - now the production code is GPU only, and the CPU code has been moved into the tests for reference purposes only. Note that the method name has been changed to add the suffix `_cupy` - not sure if that's desired?
> 
> Thanks! I think the suffix `cupy` is now redundant and can be dropped for all methods.

I removed it for this PR. We'll need to do a global refactoring to make all methods and tests consistent with this strategy after.